backend/requirements.txt
djangorestframework
PyJWT
django-cors-headers